requirements.txt
streamlit
faiss-cpu
sentence-transformers
transformers
scikit-learn
pymupdf
pandas
numpy